+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Box position to be compatible with the new header design.
+
+## [1.2.1] - 2018-08-20
+### Fixed
+- Fix email to e-mail in `pt-BR` translate.
 
 ## [1.2.1] - 2018-08-20
 ### Fixed